refactor: unify persistence modules behind one storage interface
Move the five persistence modules into src/lib/storage with a single
barrel entry point ($lib/storage):

- stores/persistence.svelte.ts -> storage/availability.svelte.ts
- utils/persistence-api.ts -> storage/api.ts (persistence-config.ts
  folded in as a private API_BASE_URL const and deleted)
- utils/session-storage.ts -> storage/working-copy.ts
- utils/load-pipeline.ts -> storage/load-pipeline.ts
- utils/persistence-manager.svelte.ts -> storage/manager.svelte.ts
  (storage half only)

Extract the non-storage app actions (fit-all, export, share, cleanup
prompt, new-rack reset) from the manager into src/lib/utils/app-actions.ts,
written without ts-nocheck. Modules inside storage/ import siblings by
relative path; production call sites import from $lib/storage and
$lib/utils/app-actions.

Delete dead exports with zero importers: persistenceStore, isCheckingApi,
recheckApiAvailability, uploadAsset, getAssetUrl. Keep deprecated
loadSession() (tests use it).

Test changes are path-only updates to vi.mock and import paths.

Refs #2027
