+++ src/lib/storage/manager.svelte.ts
@@ -4,39 +4,28 @@ import {
   setApiAvailable,
   getApiAvailableState,
   hasEverConnectedToApi,
-} from "$lib/stores/persistence.svelte";
-import {
-  saveLayoutToServer,
-  checkApiHealth,
-  PersistenceError,
-} from "$lib/utils/persistence-api";
-import { saveSession, clearSession } from "$lib/utils/session-storage";
+} from "./availability.svelte";
+import { saveLayoutToServer, checkApiHealth, PersistenceError } from "./api";
+import { saveSession, clearSession } from "./working-copy";
+import { loadFromFile } from "./load-pipeline";
 import { getLayoutStore } from "$lib/stores/layout.svelte";
-import { getUIStore } from "$lib/stores/ui.svelte";
-import { getCanvasStore } from "$lib/stores/canvas.svelte";
 import { getToastStore } from "$lib/stores/toast.svelte";
-import { getImageStore } from "$lib/stores/images.svelte";
 import { dialogStore } from "$lib/stores/dialogs.svelte";
-import { DRAWER_WIDTH } from "$lib/constants/layout";
 import { downloadYamlFile } from "$lib/utils/archive";
 import { persistenceDebug } from "$lib/utils/debug";
-import { generateShareUrl } from "$lib/utils/share";
-import { generateQRCode, canFitInQR } from "$lib/utils/qrcode";
 import {
-  generateExportSVG,
-  exportAsSVG,
-  exportAsPNG,
-  exportAsJPEG,
-  exportAsPDF,
-  exportToCSV,
-  downloadBlob,
-  generateExportFilename,
-} from "$lib/utils/export";
-import { loadFromFile } from "$lib/utils/load-pipeline";
-import type { ExportOptions } from "$lib/types";
+  shouldShowCleanupPrompt,
+  resetAndOpenNewRack,
+} from "$lib/utils/app-actions";
 
 // Internal save status (kept for circuit breaker / health check logic, not exported)
-type SaveStatusInternal = "idle" | "saving" | "saved" | "error" | "offline" | "disabled";
+type SaveStatusInternal =
+  | "idle"
+  | "saving"
+  | "saved"
+  | "error"
+  | "offline"
+  | "disabled";
 let _saveStatus = $state<SaveStatusInternal>("idle");
 
 // Circuit breaker
@@ -52,37 +41,6 @@ let saveDebounceTimer: ReturnType<typeof setTimeout> | null = null;
 
 export function getConsecutiveSaveFailures(): number {
   return _consecutiveSaveFailures;
-}
-
-export function handleFitAll(): void {
-  const layoutStore = getLayoutStore();
-  const uiStore = getUIStore();
-  const canvasStore = getCanvasStore();
-  const rightOffset = uiStore.rightDrawerOpen ? DRAWER_WIDTH : 0;
-  canvasStore.clearSavedViewport();
-  canvasStore.fitAll(layoutStore.racks, layoutStore.rack_groups, rightOffset);
-}
-
-function shouldShowCleanupPrompt(
-  operation: "save" | "saveAs" | "export",
-): boolean {
-  const uiStore = getUIStore();
-  const layoutStore = getLayoutStore();
-  if (!uiStore.promptCleanupOnSave) return false;
-  const unusedTypes = layoutStore.getUnusedCustomDeviceTypes();
-  if (unusedTypes.length === 0) return false;
-  dialogStore.pendingCleanupOperation = operation;
-  dialogStore.open("cleanupPrompt");
-  return true;
-}
-
-export function resetAndOpenNewRack(): void {
-  const layoutStore = getLayoutStore();
-  const imageStore = getImageStore();
-  layoutStore.resetLayout();
-  const usedSlugs = layoutStore.getUsedDeviceTypeSlugs();
-  imageStore.cleanupOrphanedImages(usedSlugs);
-  dialogStore.open("newRack");
 }
 
 export function handleSaveFailure(
@@ -247,156 +205,11 @@ export function maybeSaveAs(): void {
   handleSaveAsArchive();
 }
 
-export function maybeExport(): void {
-  if (shouldShowCleanupPrompt("export")) return;
-  handleExport();
-}
-
 export async function handleLoad(): Promise<void> {
   if (isApiAvailable()) {
     dialogStore.open("load");
   } else {
     await loadFromFile();
-  }
-}
-
-export async function handleExport(): Promise<void> {
-  const layoutStore = getLayoutStore();
-  const toastStore = getToastStore();
-  if (!layoutStore.hasRack) {
-    toastStore.showToast("No racks to export", "warning");
-    return;
-  }
-  try {
-    const shareUrl = generateShareUrl(layoutStore.layout);
-    if (canFitInQR(shareUrl)) {
-      dialogStore.exportQrCodeDataUrl = await generateQRCode(shareUrl, {
-        width: 444,
-      });
-    } else {
-      dialogStore.exportQrCodeDataUrl = undefined;
-    }
-  } catch {
-    dialogStore.exportQrCodeDataUrl = undefined;
-  }
-  dialogStore.open("export");
-}
-
-export function handleShare(): void {
-  const layoutStore = getLayoutStore();
-  const toastStore = getToastStore();
-  if (!layoutStore.hasRack) {
-    toastStore.showToast("No rack to share", "warning");
-    return;
-  }
-  dialogStore.open("share");
-}
-
-export async function handleExportSubmit(
-  options: ExportOptions,
-): Promise<void> {
-  const layoutStore = getLayoutStore();
-  const uiStore = getUIStore();
-  const imageStore = getImageStore();
-  const toastStore = getToastStore();
-  dialogStore.close();
-
-  try {
-    const racksToExport = options.selectedRackIds?.length
-      ? layoutStore.racks.filter((r) => options.selectedRackIds!.includes(r.id))
-      : layoutStore.racks;
-
-    if (racksToExport.length === 0) {
-      toastStore.showToast("No rack to export", "warning");
-      return;
-    }
-
-    const exportOptions = {
-      ...options,
-      displayMode: uiStore.displayMode,
-    };
-
-    const images = imageStore.getAllImages();
-    const svg = generateExportSVG(
-      racksToExport,
-      layoutStore.device_types,
-      exportOptions,
-      images,
-      layoutStore.rack_groups,
-    );
-
-    const exportViewOrDefault = options.exportView ?? "both";
-
-    const imageFormatHandlers: Record<
-      string,
-      (
-        svg: SVGSVGElement,
-        layoutName: string,
-        exportView: string,
-      ) => Promise<void>
-    > = {
-      svg: async (svgEl, layoutName, exportView) => {
-        const svgString = exportAsSVG(svgEl);
-        const blob = new Blob([svgString], { type: "image/svg+xml" });
-        downloadBlob(
-          blob,
-          generateExportFilename(layoutName, exportView, "svg"),
-        );
-        toastStore.showToast("SVG exported successfully", "success");
-      },
-      png: async (svgEl, layoutName, exportView) => {
-        const blob = await exportAsPNG(svgEl);
-        downloadBlob(
-          blob,
-          generateExportFilename(layoutName, exportView, "png"),
-        );
-        toastStore.showToast("PNG exported successfully", "success");
-      },
-      jpeg: async (svgEl, layoutName, exportView) => {
-        const blob = await exportAsJPEG(svgEl);
-        downloadBlob(
-          blob,
-          generateExportFilename(layoutName, exportView, "jpeg"),
-        );
-        toastStore.showToast("JPEG exported successfully", "success");
-      },
-      pdf: async (svgEl, layoutName, exportView) => {
-        const svgString = exportAsSVG(svgEl);
-        const blob = await exportAsPDF(svgString, options.background);
-        downloadBlob(
-          blob,
-          generateExportFilename(layoutName, exportView, "pdf"),
-        );
-        toastStore.showToast("PDF exported successfully", "success");
-      },
-    };
-
-    const handler = imageFormatHandlers[options.format];
-    if (handler) {
-      await handler(svg, layoutStore.layout.name, exportViewOrDefault);
-    } else if (options.format === "csv") {
-      const firstRack = racksToExport[0];
-      if (!firstRack) {
-        throw new Error("No rack available for CSV export");
-      }
-      const csvContent = exportToCSV(firstRack, layoutStore.device_types);
-      const blob = new Blob([csvContent], { type: "text/csv;charset=utf-8" });
-      downloadBlob(
-        blob,
-        generateExportFilename(layoutStore.layout.name, null, "csv"),
-      );
-      const successMsg =
-        racksToExport.length > 1
-          ? `CSV exported (first rack only - "${firstRack.name}")`
-          : "CSV exported successfully";
-      toastStore.showToast(successMsg, "success");
-    }
-  } catch (error) {
-    persistenceDebug.api("Export failed: %O", error);
-    toastStore.showToast(
-      error instanceof Error ? error.message : "Export failed",
-      "error",
-    );
   }
 }
